re PR middle-end/90478 (ICE in emit_case_dispatch_table at gcc/stmt.c:796)
	PR middle-end/90478
	* gcc.dg/tree-ssa/pr90478.c: Add empty dg-options.  Use long long type
	instead of long.

From-SVN: r271271

+++ gcc/testsuite/gcc.dg/tree-ssa/pr90478.c
@@ -1,7 +1,8 @@
 /* { dg-do compile } */
+/* { dg-options "" } */
 
 typedef struct {
-  long a;
+  long long a;
 } c;
 
 void e();
